// src/pages/AdminPage/components/Processes/ProcessEditorPanel.module.scss.d.ts
declare const styles: {
  editorOverlay: string;
  editorPanel: string;
  editorHeader: string;
  editorDescription: string;
  closeButton: string;
  editorForm: string;
  formGrid: string;
  formGroup: string;
  inputControl: string;
  checkboxGroup: string;
  checkboxControl: string;
  editorActions: string;
  primaryButton: string;
  secondaryButton: string;
};

export default styles;